--- src/manager/module_inside/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ModuleInsideConfig(AppConfig):
-    name = 'module_inside'
