--- src/manager/module_inside/proto/tea.py
@@ -1,50 +1,0 @@
-"""
-TencentBlueKing is pleased to support the open source community by making
-蓝鲸智云PaaS平台社区版 (BlueKing PaaSCommunity Edition) available.
-Copyright (C) 2017-2018 THL A29 Limited,
-a Tencent company. All rights reserved.
-Licensed under the MIT License (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at http://opensource.org/licenses/MIT
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on
-an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
-either express or implied. See the License for the
-specific language governing permissions and limitations under the License.
-"""
-
-from drf_yasg.utils import swagger_auto_schema
-from rest_framework import serializers
-
-from common.drf.serializers import BaseRspSerializer
-from src.manager.module_inside.proto import inside_tag
-
-
-class ReqGetProductSerializer(serializers.Serializer):
-    """
-    tea查询
-    """
-
-    cc_id = serializers.CharField(label="cc_id", required=False)
-
-
-class RspGetProductSerializer(BaseRspSerializer):
-    class RspGetProductSerializerData(serializers.Serializer):
-        class RspGetProductSerializerDataResult(serializers.Serializer):
-            name = serializers.CharField(
-                label="业务名称",
-            )
-            code = serializers.CharField(label="唯一编码")
-        result = serializers.ListField(child=RspGetProductSerializerDataResult())
-        count = serializers.IntegerField(label="业务数量")
-    data = RspGetProductSerializerData()
-
-
-
-############################################################
-get_product_docs = swagger_auto_schema(
-    tags=inside_tag,
-    operation_id="查询tea业务",
-    query_serializer=ReqGetProductSerializer(),
-    responses={200: RspGetProductSerializer()},
-)
